Modify: follow check route주소 변경 - 기존 works/feed/:id/followcheck - 변경 follow/:id
