+++ controllers/workController.js
@@ -1,13 +1,5 @@
 const workService = require('../services/workService');
 
-// 상세피드에서 로그인유저의 팔로잉 체크여부
-const followCheck = async (req, res) => {
-  const { id } = req.params;
-  const user_id = req.user_id;
-  const result = await workService.followCheck(id, user_id);
-  res.status(200).json(result);
-};
-
 // 카테고리별 총 게시물 수 + 최신 feed list
 const worksList = async (req, res) => {
   const { sort } = req.query;
@@ -23,7 +15,6 @@ const feed = async (req, res) => {
 };
 
 module.exports = {
-  followCheck,
   worksList,
   feed,
 };
