+++ models/workDao.js
@@ -88,34 +88,6 @@ const worksList = async (isSelect, sortOfOrder) => {
   return { categorySortCountList, worksFeedList };
 };
 
-// feed 글쓴이와 유저와의 팔로우 관계
-const followCheck = async (id, user_id) => {
-  return await myDataSource
-    .query(
-      `
-      SELECT
-        EXISTS (
-          SELECT
-            f.id
-          FROM
-            Follow f
-          LEFT JOIN Works_Posting wp ON
-            wp.user_id = f.following_id
-          WHERE
-            wp.id = ?
-            AND follower_id = ?
-        ) AS success
-    `,
-      [id, user_id]
-    )
-    .then(value => {
-      const [item] = value;
-      return {
-        follow_check: item.success == 1,
-      };
-    });
-};
-
 // feed 상세
 const feed = async id => {
   // feed img_url 배열(다수의 이미지가 있을 시)
@@ -519,7 +491,6 @@ const feed = async id => {
 };
 
 module.exports = {
-  followCheck,
   worksList,
   feed,
 };
