+++ services/followService.js
@@ -1,5 +1,10 @@
 const followDao = require('../models/followDao');
 
+// 상세피드에서 팔로우 체결유무 확인
+const followCheck = async (id, user_id) => {
+  return await followDao.followCheck(id, user_id);
+};
+
 // 팔로우 체결
 // TODO 팔로우 체크 도입해서 중복팔로우 되지 않게
 const following = async (following_id, user_id) => {
@@ -12,6 +17,7 @@ const followingCancel = async (following_id, user_id) => {
 };
 
 module.exports = {
+  followCheck,
   following,
   followingCancel,
 };
